Refactoring. Preparation for "clone" and "wi".
Each command is a self-sufficient executable that can be ran without
git-tf. It is done because clone and wi do not need the "init" in
main.py.

Now each command is accessible as an executable with name "git-tf-<command>".
It is a pattern followed by git: any git command is accessible as an executable in
  /usr/lib/git-core.

A git-tf-<command> executable is just a symlink to a <command>.py files.