Numer of reported TFS commits for status and log fixed.
When computing the commits to be pushed on TFS the following command
is executed:

commits = git('log %s.. --format=%%H master --reverse --first-parent'
              % lastCommit).splitlines()

But for "git tf log" and "git tf status" there was no --first-parent
option. Because of this the number of commits reported with "tf status"
and "tf log" consisted of commits in the current branch and commits from
all merged branches (event when the merge was not fast-forward). The
actual TFS commits are only the ones from the current branch.